+++ blocks/dashboard/dashboard.css
@@ -199,92 +199,6 @@ dialog .admin img {
     padding-bottom: 12px;
 }
 
-.dashboard.block .analytics {
-    margin-bottom: 32px;
-}
-
-.dashboard.block .analytics .title {
-    padding-bottom: 20px;
-}
-
-.dashboard.block .analytics .title h2 {
-    padding: 0;
-}
-
-.dashboard.block .analytics .chart-container {
-    position: relative;
-    height: 400px;
-    width: 100%;
-    margin-block: 32px;
-}
-
-.dashboard.block .analytics .cards {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 32px;
-    align-items: stretch;
-    justify-content: center;
-}
-
-.dashboard.block .analytics .cards .box {
-    display: flex;
-    font-weight: 300;
-    text-overflow: ellipsis;
-    overflow: hidden;
-    flex: 1 1 300px;
-    min-width: 200px;
-    max-width: 300px;
-    width: 100%;
-    box-sizing: border-box;
-    text-align: center;
-    word-wrap: break-word;
-    flex-direction: column;
-    gap: 8px;
-    border: 1px solid var(--gray);
-    padding: 16px;
-    border-radius: var(--border-radius);
-}
-
-.dashboard.block .analytics .cards .box strong {
-    text-transform: uppercase;
-}
-
-.dashboard.block .analytics .cards .box strong:only-child::after {
-    content: ' (Not enough data)';
-    display: block;
-    font-size: 11px;
-}
-
-.dashboard.block .analytics .cards .box p {
-    display: flex;
-    justify-content: space-between;
-    margin: 0;
-}
-
-.dashboard.block .analytics .cards .box p span:first-child {
-    white-space: nowrap;
-    overflow: hidden;
-    margin-right: 8px;
-    text-overflow: ellipsis;
-}
-
-@media (width <= 1200px) {
-    .dashboard.block .analytics .cards .box {
-        flex: 1 1 calc(50% - 32px) !important;
-    }
-}
-
-@media (width < 768px) {
-    .dashboard.block .analytics .cards {
-        flex-direction: column;
-    }
-
-    .dashboard.block .analytics .cards .box {
-        flex: 1 1 100% !important;
-        max-width: 100% !important;
-    }
-}
-
 body:not(.is-headwire, .is-test-user) .dashboard.block .admin {
     display: none;
 }
