+++ styles/analytics.css
@@ -1,0 +1,60 @@
+.analytics {
+    margin-bottom: 32px;
+}
+
+.analytics .title {
+    padding-bottom: 20px;
+}
+
+.analytics .title h2 {
+    padding: 0;
+}
+
+.analytics .chart-container {
+    position: relative;
+    height: 400px;
+    width: 100%;
+    margin-block: 32px;
+}
+
+.analytics .cards .cwp-box {
+    flex: 1 1 calc(25% - 32px);
+    min-width: 200px;
+    max-width: 300px;
+    box-sizing: border-box;
+    text-align: center;
+    word-wrap: break-word;
+}
+
+.analytics .cards .cwp-box ul {
+    list-style: none;
+    padding: 0 !important;
+    display: flex;
+    flex-direction: column;
+    gap: 8px;
+}
+
+.analytics .cards.metrics span {
+    display: flex;
+    justify-content: center;
+}
+
+.analytics .cards.metrics ul {
+    margin: 0;
+    font-weight: 300;
+    padding-left: 16px;
+}
+
+.analytics .cards.metrics > div > strong:only-child::after {
+    content: ' (Not enough data)';
+    display: block;
+    font-size: 11px;
+}
+
+.analytics .cards .red {
+    color: var(--red);
+}
+
+.analytics .cards .green {
+    color: var(--green);
+}
